backend/ai_models/requirements.txt
pillow>=9.0.0
numpy>=1.21.0
tensorflow>=2.6.0